# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 智岗罗盘 API
  version: 1.0.0
  description: |
    多源异构驱动的岗位能力动态演化与人岗匹配系统 API 契约。
    本文件是前后端并行开发的单一事实源（AGENTS.md 铁律一）。
    前端通过 `openapi-typescript` 生成类型，不得自行定义与后端不一致的类型。
    认证方案：Bearer 双 Token（access 30min / refresh 7d）；refresh_token 同时
    写入 httpOnly Cookie（httponly + samesite=lax），供页面刷新后静默恢复会话。
    标注规则：全部端点已实现（M1-M4 全量交付），契约不含里程碑占位标注。
  contact:
    name: 智岗罗盘团队
    url: https://github.com/zhigang-compass
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.zhigang.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/match/recommend:
    post:
      tags: [match]
      summary: 自动推荐 Top-N（异步，ARQ 队列 + 含 CII 通胀修正）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resume_id]
              properties:
                resume_id: { type: string }
                top_n: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        '202':
          description: 任务已入队，返回 task_id（轮询 /match/task/{task_id} 获取进度与结果）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/RecommendTaskResult'
                  trace_id: { type: string }
        '422':
          description: resume_id 格式非法（code 4000）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '403':
          description: 非本人无权使用该简历发起匹配
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '404':
          description: 简历不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/v1/match/compare:
    post:
      tags: [match]
      summary: 人岗比对（含差距三态 + 学习路径）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resume_id, position_id]
              properties:
                resume_id: { type: string }
                position_id: { type: string }
      responses:
        '200':
          description: |
            匹配结果 + gaps（GapSkill[]，三态）+ learning_path
            （LearningPathItem[]，missing/weak 技能先修链 + 课程 Top-3）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/MatchResult'
                  trace_id: { type: string }
  /api/v1/match/task/{task_id}:
    get:
      tags: [match]
      summary: 查询推荐任务状态
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 任务状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/MatchTaskStatus'
                  trace_id: { type: string }
  /api/v1/match/result/{match_id}:
    get:
      tags: [match]
      summary: 获取匹配结果
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: 匹配详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/MatchResultList'
                  trace_id: { type: string }
  /api/v1/match/result/{match_id}/jd/{jd_id}:
    get:
      tags: [match]
      summary: 获取比对岗位下某条 JD 的详情（下拉逐条查看各 JD 评分/差距/路径）
      parameters:
        - $ref: '#/components/parameters/Id'
        - in: path
          name: jd_id
          required: true
          schema: { type: string }
          description: 原生 JD 主键 id（来自 compare 响应 jd_breakdown[].jd_id）
      responses:
        '200':
          description: 单 JD 匹配详情（并入当前岗位详情展示；岗位级字段 domain_name/weights 不变）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data: { $ref: '#/components/schemas/MatchResult' }
                  trace_id: { type: string }
  /api/v1/match/result/{match_id}/gap:
    get:
      tags: [match]
      summary: 获取差距分析
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: 差距列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/MatchGapData'
                  trace_id: { type: string }
  /api/v1/match/result/{match_id}/path:
    get:
      tags: [match]
      summary: 获取学习路径
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: 学习路径
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/MatchPathData'
                  trace_id: { type: string }
  /api/v1/match/result/{match_id}/diagnosis:
    post:
      tags: [match]
      summary: 异步生成或复用人岗比对诊断报告
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '202':
          description: 诊断任务已创建或复用
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/DiagnosisTaskResponse'
                  trace_id: { type: string }
        '422':
          description: 该匹配结果无差距数据（code 4000）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '404':
          description: 匹配结果不存在或无权访问
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    get:
      tags: [match]
      summary: 获取人岗比对诊断报告（只读）
      description: >
        读取已生成的诊断报告：24h Redis 缓存优先，过期后从 PostgreSQL
        DiagnosisReportRecord 耐久记录回读并回填缓存。生成一律走 POST
        异步任务（worker 唯一执行路径），本端点不再同步调用 LLM；
        缓存与落库皆无时返回 404，客户端应先 POST 创建生成任务。
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: 诊断报告（overall_summary / radar_analysis / top_gaps / path_analysis / recommendations）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/DiagnosisReport'
                  trace_id: { type: string }
        '422':
          description: 该匹配结果无差距数据（仅人岗比对可生成诊断报告，code 4000）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '404':
          description: 匹配结果不存在或已过期；或诊断报告尚未生成（需先 POST 创建生成任务）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/v1/match/feedback:
    post:
      tags: [match]
      summary: 提交匹配反馈
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [match_id, score]
              properties:
                match_id: { type: string }
                score: { type: integer, enum: [1, -1], description: "1=👍, -1=👎" }
      responses:
        '200':
          description: 反馈已记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================
  # 演化模块 /api/v1/evolution
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    ApiResponse:
      type: object
      required: [code, msg, trace_id]
      properties:
        code:
          type: integer
          description: |
            0=成功，非 0=业务错误码（设计文档 §2.4.7）：4000 参数校验 / 4010 未认证 /
            4011 Token 过期 / 4030 无权限 / 4040 资源不存在 / 4090 资源或状态冲突 /
            4290 请求限流 / 5000 内部错误（含 LLM 能力不可用场景，HTTP 503）/
            5001 Neo4j 查询失败 / 5002 pgvector 查询失败 /
            5003 LLM 超时（HTTP 504）
        msg:
          type: string
        data:
          description: 业务数据（成功时存在）
        trace_id:
          type: string
          description: 链路追踪 ID（uuid4 hex 截断 16 位）
...
    MatchTaskStatus:
      type: object
      description: 匹配任务状态（GET /match/task/{task_id}）
      required: [task_id, status]
      properties:
        task_id: { type: string }
        status: { type: string, enum: [pending, running, success, failed] }
        progress: { type: integer, description: "进度 [0,100]" }
        error: { type: string }
        match_id: { type: string, description: 成功后附，供拉取结果 }
    DiagnosisTaskResponse:
      type: object
      description: POST /match/result/{match_id}/diagnosis 响应 data
      required: [task_id, status]
      properties:
        task_id: { type: string }
        status: { type: string, enum: [pending, running, success, failed] }
        match_id: { type: string }
        report: { $ref: '#/components/schemas/DiagnosisReport', nullable: true }
        error: { type: string }
    RecommendTaskResult:
      type: object
      description: POST /match/recommend 响应 data（ARQ 入队）
      required: [task_id]
      properties:
        task_id: { type: string }
    MatchResultList:
      type: object
      description: GET /match/result/{match_id} 响应 data
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/MatchResult' }
          description: 推荐任务结果列表（recommend 多岗位）
    MatchGapData:
      type: object
      description: GET /match/result/{match_id}/gap 响应 data
      required: [match_id, gaps]
      properties:
        match_id: { type: string }
        gaps:
          type: array
          items: { $ref: '#/components/schemas/GapSkill' }
    MatchPathData:
      type: object
      description: GET /match/result/{match_id}/path 响应 data
      required: [match_id, learning_path]
      properties:
        match_id: { type: string }
        learning_path:
          type: array
          items: { $ref: '#/components/schemas/LearningPathItem' }
        learning_path_blocked:
          type: boolean
          default: false
          description: 学习路径是否因领域跨簇语义黑名单拦截（P1：跨域诱导组合拒绝生成）
        learning_path_block_reason:
          type: string
          nullable: true
          default: null
          description: 拦截原因（命中的岗位行业 × 候选人领域对），未拦截为 null
    GapAdvice:
      type: object
      description: 诊断报告差距建议项（evidence_id 可点击追溯）
      required: [skill, advice, evidence_id]
      properties:
        skill: { type: string }
        advice: { type: string }
        evidence_id: { type: string }
    DiagnosisReport:
      type: object
      description: LLM 人岗比对诊断报告（GET /match/result/{match_id}/diagnosis，§9.5）
      required: [overall_summary, radar_analysis, top_gaps, recommendations]
      properties:
        match_id: { type: string }
        overall_summary: { type: string, description: 总体匹配度解读 }
        radar_analysis: { type: string, description: 三维雷达图解读 }
        top_gaps:
          type: array
          items: { $ref: '#/components/schemas/GapAdvice' }
          description: 关键差距 Top-5 及改进建议
        path_analysis: { type: string, description: 学习路径解读 }
        recommendations:
          type: array
          items: { type: string }
          description: 整体改进建议清单

    # ============================================================
    # 图算法端点 data（08-14 契约补全：pagerank/skill-clusters/shortest-path）
    # ============================================================
...